LayoutTests/media/track/text-track-in-band-exposure.html

<!doctype html>
<html>
<head>
    <title>In-band text tracks: exposure, readiness, dispatch type</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../media-source/mp4-generator.js"></script>
    <script src="caption-generator.js"></script>
</head>
<body>
<div id="log"></div>
<script>
// HTML Living Standard:
//   §4.8.13.7 "Text tracks":
//     - "in-band metadata track dispatch type" definition
//       https://html.spec.whatwg.org/multipage/media.html#text-track-in-band-metadata-track-dispatch-type
//   §4.8.13.7.1 "Sourcing in-band text tracks":
//     https://html.spec.whatwg.org/multipage/media.html#sourcing-in-band-text-tracks
//   §4.8.13.10 "Text track API":
//     - TextTrack.inBandMetadataTrackDispatchType
//       https://html.spec.whatwg.org/multipage/media.html#dom-texttrack-inbandmetadatatrackdispatchtype
//
// In-band text tracks (those exposed by the media resource itself, not via
// a <track> element) must:
//   - have readiness state "loaded" by the time they are exposed
//   - have an empty inBandMetadataTrackDispatchType for non-metadata kinds
//   - be removed from the media element's list of pending text tracks
//     during load (i.e. they must not block readyState progression)
//
// We exercise this with synthetic CEA-608 captions injected via
// mp4-generator.js + caption-generator.js — these produce a real fragmented
// MP4 with an in-band CEA-608 captions track, exercising the platform's
// in-band parser path.
//
// TODO: the metadata-track case (where inBandMetadataTrackDispatchType
// must be non-empty per §4.8.13.7.1 step 6) requires extending
// mp4-generator.js to emit a 'mett'/'metx' sample-entry track. That's
// deferred to a future round.

function buildCaptionedMp4() {
    const FRAME_COUNT = 60;
    const FRAME_DURATION = 1001;     // 30000/1001 ≈ 29.97fps
    const TIMESCALE = 30000;
    const TOTAL_TICKS = FRAME_COUNT * FRAME_DURATION;
    const CUE_TICKS = 20 * FRAME_DURATION;
    const CAPTION_START_OFFSET = TIMESCALE; // 1 second

    const captionSamples = Captions.buildPopOnCueSamples({
        cues: [
            { text: 'Cue 1', durationTicks: CUE_TICKS },
            { text: 'Cue 2', durationTicks: CUE_TICKS },
        ],
        totalDurationTicks: TOTAL_TICKS,
    });

    const videoSamples = Array.from({ length: FRAME_COUNT }, (_, i) => ({
        duration: FRAME_DURATION,
        isSync: i % 30 === 0,
        dts: i * FRAME_DURATION,
        pts: i * FRAME_DURATION,
    }));

    const { init, media } = MP4.samples({
        timescale: TIMESCALE,
        tracks: [
            { id: 1, type: 'video' },
            { id: 2, type: 'captions' },
        ],
        trackSamples: {
            1: videoSamples,
            2: captionSamples.map(s => ({
                duration: s.duration,
                data: s.data,
                isSync: true,
                dts: CAPTION_START_OFFSET,
                pts: CAPTION_START_OFFSET,
            })),
        },
    });

    const combined = new Uint8Array(init.byteLength + media.byteLength);
    combined.set(new Uint8Array(init), 0);
    combined.set(new Uint8Array(media), init.byteLength);
    return URL.createObjectURL(new Blob([combined], { type: 'video/mp4' }));
}

function makeVideo(t) {
    const video = document.createElement("video");
    video.muted = true;
    document.body.appendChild(video);
    t.add_cleanup(() => video.remove());
    return video;
}

test(() => {
    // The IDL attribute exists on TextTrack and is a string-typed accessor.
    const dummy = document.createElement("video").addTextTrack("captions");
    assert_true("inBandMetadataTrackDispatchType" in dummy,
        "TextTrack exposes inBandMetadataTrackDispatchType");
    assert_equals(typeof dummy.inBandMetadataTrackDispatchType, "string");
    // For an addTextTrack-created (non-in-band) track it must be empty.
    assert_equals(dummy.inBandMetadataTrackDispatchType, "",
        "addTextTrack-created track has empty inBandMetadataTrackDispatchType");
}, "TextTrack.inBandMetadataTrackDispatchType: IDL attribute exists, empty for non-in-band tracks");

promise_test(async t => {
    // Synthetic captioned MP4 -> in-band track exposed via addtrack event.
    const video = makeVideo(t);
    video.src = buildCaptionedMp4();

    const ev = await new Promise((resolve, reject) => {
        video.textTracks.addEventListener("addtrack", resolve, { once: true });
        video.addEventListener("error", reject, { once: true });
        t.step_timeout(() => reject(new Error("addtrack timeout")), 5000);
    });

    assert_equals(ev.track, video.textTracks[0],
        "addtrack event fires with the in-band TextTrack");
    assert_equals(ev.track.kind, "captions",
        "in-band CEA-608 track kind is 'captions'");
}, "in-band track is exposed via addtrack on the media element");

promise_test(async t => {
    // Behavior under §4.8.13.7.1 sourcing-in-band-text-tracks step 6:
    // For a non-metadata in-band track, the dispatch type is the empty
    // string. (Metadata-kind tracks are deferred to a future round; see
    // file header.)
    const video = makeVideo(t);
    video.src = buildCaptionedMp4();

    await new Promise((resolve, reject) => {
        video.textTracks.addEventListener("addtrack", resolve, { once: true });
        video.addEventListener("error", reject, { once: true });
        t.step_timeout(() => reject(new Error("addtrack timeout")), 5000);
    });

    const tt = video.textTracks[0];
    assert_not_equals(tt.kind, "metadata",
        "synthetic CEA-608 track is NOT metadata-kind");
    assert_equals(tt.inBandMetadataTrackDispatchType, "",
        "non-metadata in-band track has empty inBandMetadataTrackDispatchType");
}, "in-band non-metadata track: inBandMetadataTrackDispatchType is empty");

promise_test(async t => {
    // §4.8.13.7.1 step 5: "Set the new text track's readiness state to
    // loaded." In-band tracks must NEVER be in pending-text-tracks state;
    // their readiness is loaded by the time they're exposed. We verify
    // indirectly: the video reaches HAVE_FUTURE_DATA (canplay) without
    // ever being blocked on the in-band track.
    const video = makeVideo(t);
    video.src = buildCaptionedMp4();

    let addtrackFired = false;
    video.textTracks.addEventListener("addtrack",
        () => { addtrackFired = true; }, { once: true });

    await new Promise((resolve, reject) => {
        video.addEventListener("canplay", resolve, { once: true });
        video.addEventListener("error", reject, { once: true });
        t.step_timeout(() => reject(new Error("canplay timeout")), 5000);
    });

    assert_true(addtrackFired,
        "addtrack must fire by the time canplay reaches the page");
    assert_greater_than_equal(video.readyState, video.HAVE_FUTURE_DATA,
        "video reaches HAVE_FUTURE_DATA -- in-band track did not block readyState");
}, "in-band track does not block readyState progression (loaded immediately)");
</script>
</body>
</html>
